requirements.txt
typer
InquirerPy
deepeval
python-dotenv
pandas 
numpy